requirements.txt
deepl
git+https://github.com/Rapptz/discord.py
python-dotenv
lingua-language-detector
pymongo
